requirements.txt
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
kubernetes==29.0.0
kopf==1.36.2
litellm==1.17.0
fastapi==0.109.0
uvicorn==0.27.0
PyYAML==6.0.1
chromadb==0.4.22
pandas==2.2.0
numpy==1.26.3
